Created autocodesign package, moved autoCodesign function
