+++ main.go
@@ -1,246 +1,20 @@
 package main
 
 import (
-	"context"
 	"fmt"
-	"io/ioutil"
-	"net/http"
-	"net/url"
 	"os"
-	"strings"
-	"time"
 
 	"github.com/bitrise-io/go-steputils/stepconf"
 	"github.com/bitrise-io/go-steputils/tools"
 	"github.com/bitrise-io/go-utils/log"
-	"github.com/bitrise-io/go-utils/retry"
 	"github.com/bitrise-io/go-xcode/appleauth"
-	"github.com/bitrise-io/go-xcode/certificateutil"
-	"github.com/bitrise-io/go-xcode/devportalservice"
-	"github.com/bitrise-steplib/steps-ios-auto-provision-appstoreconnect/appstoreconnect"
+	"github.com/bitrise-steplib/steps-ios-auto-provision-appstoreconnect/autocodesign"
 	"github.com/bitrise-steplib/steps-ios-auto-provision-appstoreconnect/autoprovision"
-	"github.com/bitrise-steplib/steps-ios-auto-provision-appstoreconnect/spaceship"
 )
-
-// downloadCertificates downloads and parses a list of p12 files
-func downloadCertificates(URLs []CertificateFileURL) ([]certificateutil.CertificateInfoModel, error) {
-	httpClient := &http.Client{
-		Timeout: 10 * time.Second,
-	}
-	var certInfos []certificateutil.CertificateInfoModel
-
-	for i, p12 := range URLs {
-		log.Debugf("Downloading p12 file number %d from %s", i, p12.URL)
-
-		p12CertInfos, err := downloadPKCS12(httpClient, p12.URL, p12.Passphrase)
-		if err != nil {
-			return nil, err
-		}
-		log.Debugf("Codesign identities included:\n%s", autoprovision.CertsToString(p12CertInfos))
-
-		certInfos = append(certInfos, p12CertInfos...)
-	}
-
-	return certInfos, nil
-}
-
-// downloadPKCS12 downloads a pkcs12 format file and parses certificates and matching private keys.
-func downloadPKCS12(httpClient *http.Client, certificateURL, passphrase string) ([]certificateutil.CertificateInfoModel, error) {
-	contents, err := downloadFile(httpClient, certificateURL)
-	if err != nil {
-		return nil, err
-	} else if contents == nil {
-		return nil, fmt.Errorf("certificate (%s) is empty", certificateURL)
-	}
-
-	infos, err := certificateutil.CertificatesFromPKCS12Content(contents, passphrase)
-	if err != nil {
-		return nil, fmt.Errorf("failed to parse certificate (%s), err: %s", certificateURL, err)
-	}
-
-	return infos, nil
-}
-
-func downloadFile(httpClient *http.Client, src string) ([]byte, error) {
-	url, err := url.Parse(src)
-	if err != nil {
-		return nil, fmt.Errorf("failed to parse url (%s): %s", src, err)
-	}
-
-	// Local file
-	if url.Scheme == "file" {
-		src := strings.Replace(src, url.Scheme+"://", "", -1)
-
-		return ioutil.ReadFile(src)
-	}
-
-	// Remote file
-	req, err := http.NewRequest(http.MethodGet, url.String(), nil)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create request: %s", err)
-	}
-
-	var contents []byte
-	err = retry.Times(2).Wait(5 * time.Second).Try(func(attempt uint) error {
-		log.Debugf("Downloading %s, attempt %d", src, attempt)
-
-		ctx, cancel := context.WithTimeout(context.Background(), 1*time.Minute)
-		defer cancel()
-		req = req.WithContext(ctx)
-
-		resp, err := httpClient.Do(req)
-		if err != nil {
-			return fmt.Errorf("failed to download (%s): %s", src, err)
-		}
-		defer func() {
-			if err := resp.Body.Close(); err != nil {
-				log.Warnf("failed to close (%s) body: %s", src, err)
-			}
-		}()
-
-		if resp.StatusCode != http.StatusOK {
-			return fmt.Errorf("download (%s) failed with status code (%d)", src, resp.StatusCode)
-		}
-
-		contents, err = ioutil.ReadAll(resp.Body)
-		if err != nil {
-			return fmt.Errorf("failed to read response (%s): %s", src, err)
-		}
-
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return contents, nil
-}
 
 func failf(format string, args ...interface{}) {
 	log.Errorf(format, args...)
 	os.Exit(1)
-}
-
-const notConnected = `Bitrise Apple service connection not found.
-Most likely because there is no configured Bitrise Apple service connection.
-Read more: https://devcenter.bitrise.io/getting-started/configuring-bitrise-steps-that-require-apple-developer-account-data/`
-
-func autoCodesign(buildURL, buildAPIToken string,
-	authSources []appleauth.Source, certificateURLs []CertificateFileURL, distributionType autoprovision.DistributionType,
-	signUITestTargets, verboseLog bool,
-	codesignRequirements autoprovision.CodesignRequirements, minProfileDaysValid int,
-	keychainPath string, keychainPassword stepconf.Secret) (map[autoprovision.DistributionType]autoprovision.CodesignSettings, error) {
-
-	fmt.Println()
-	log.Infof("Fetching Apple service connection")
-	connectionProvider := devportalservice.NewBitriseClient(retry.NewHTTPClient().StandardClient(), buildURL, buildAPIToken)
-	conn, err := connectionProvider.GetAppleDeveloperConnection()
-	if err != nil {
-		if networkErr, ok := err.(devportalservice.NetworkError); ok && networkErr.Status == http.StatusUnauthorized {
-			fmt.Println()
-			log.Warnf("Unauthorized to query Bitrise Apple service connection. This happens by design, with a public app's PR build, to protect secrets.")
-			return nil, err
-		}
-
-		fmt.Println()
-		log.Errorf("Failed to activate Bitrise Apple service connection")
-		log.Warnf("Read more: https://devcenter.bitrise.io/getting-started/configuring-bitrise-steps-that-require-apple-developer-account-data/")
-
-		return nil, err
-	}
-
-	if len(conn.DuplicatedTestDevices) != 0 {
-		log.Debugf("Devices with duplicated UDID are registered on Bitrise, will be ignored:")
-		for _, d := range conn.DuplicatedTestDevices {
-			log.Debugf("- %s, %s, UDID (%s), added at %s", d.Title, d.DeviceType, d.DeviceID, d.UpdatedAt)
-		}
-	}
-
-	authConfig, err := appleauth.Select(conn, authSources, appleauth.Inputs{})
-	if err != nil {
-		if conn.APIKeyConnection == nil && conn.AppleIDConnection == nil {
-			fmt.Println()
-			log.Warnf("%s", notConnected)
-		}
-		return nil, fmt.Errorf("could not configure Apple service authentication: %v", err)
-	}
-
-	if authConfig.APIKey != nil {
-		log.Donef("Using Apple service connection with API key.")
-	} else if authConfig.AppleID != nil {
-		log.Donef("Using Apple service connection with Apple ID.")
-	} else {
-		panic("No Apple authentication credentials found.")
-	}
-
-	// create developer portal client
-	fmt.Println()
-	log.Infof("Initializing Developer Portal client")
-	var devportalClient autoprovision.DevportalClient
-	if authConfig.APIKey != nil {
-		httpClient := appstoreconnect.NewRetryableHTTPClient()
-		client := appstoreconnect.NewClient(httpClient, authConfig.APIKey.KeyID, authConfig.APIKey.IssuerID, []byte(authConfig.APIKey.PrivateKey))
-		client.EnableDebugLogs = false // Turn off client debug logs including HTTP call debug logs
-		devportalClient = autoprovision.NewAPIDevportalClient(client)
-		log.Donef("App Store Connect API client created with base URL: %s", client.BaseURL)
-	} else if authConfig.AppleID != nil {
-		client, err := spaceship.NewClient(*authConfig.AppleID, codesignRequirements.TeamID)
-		if err != nil {
-			return nil, fmt.Errorf("failed to initialize Apple ID client: %v", err)
-		}
-		devportalClient = spaceship.NewSpaceshipDevportalClient(client)
-		log.Donef("Apple ID client created")
-	}
-
-	// Downloading certificates
-	fmt.Println()
-	log.Infof("Downloading certificates")
-
-	certs, err := downloadCertificates(certificateURLs)
-	if err != nil {
-		return nil, fmt.Errorf("Failed to download certificates: %s", err)
-	}
-
-	log.Printf("%d certificates downloaded:", len(certs))
-
-	for _, cert := range certs {
-		log.Printf("- %s", cert.CommonName)
-	}
-
-	certsByType, distrTypes, err := autoprovision.SelectCertificatesAndDistributionTypes(
-		devportalClient.CertificateSource,
-		certs,
-		distributionType,
-		codesignRequirements.TeamID,
-		signUITestTargets,
-		verboseLog,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("%v", err)
-	}
-
-	// Ensure devices
-	var devPortalDeviceIDs []string
-	if autoprovision.DistributionTypeRequiresDeviceList(distrTypes) {
-		var err error
-		devPortalDeviceIDs, err = autoprovision.EnsureTestDevices(devportalClient.DeviceClient, conn.TestDevices, codesignRequirements.Platform)
-		if err != nil {
-			return nil, fmt.Errorf("Failed to ensure test devices: %s", err)
-		}
-	}
-
-	// Ensure Profiles
-	codesignSettingsByDistributionType, err := autoprovision.EnsureProfiles(devportalClient.ProfileClient, distrTypes, certsByType, codesignRequirements, devPortalDeviceIDs, minProfileDaysValid)
-	if err != nil {
-		return nil, fmt.Errorf("Failed to ensure profiles: %s", err)
-	}
-
-	// Install certificates and profiles
-	if err := autoprovision.InstallCertificatesAndProfiles(codesignSettingsByDistributionType, keychainPath, keychainPassword); err != nil {
-		return nil, fmt.Errorf("Failed to install codesigning files: %s", err)
-	}
-
-	return codesignSettingsByDistributionType, nil
 }
 
 func main() {
@@ -282,7 +56,7 @@ func main() {
 		failf("%v", err)
 	}
 
-	codesignSettingsByDistributionType, err := autoCodesign(stepConf.BuildURL, stepConf.BuildAPIToken, authSources, certURLs, stepConf.DistributionType(), stepConf.SignUITestTargets,
+	codesignSettingsByDistributionType, err := autocodesign.Do(stepConf.BuildURL, stepConf.BuildAPIToken, authSources, certURLs, stepConf.DistributionType(), stepConf.SignUITestTargets,
 		stepConf.VerboseLog, codesignRequirements, stepConf.MinProfileDaysValid, stepConf.KeychainPath, stepConf.KeychainPassword)
 	if err != nil {
 		failf("Automatic code signing failed: %s", err)
